@@ -14,37 +14,35 @@
     }
   ],
   "require": {
-    "php": "^7.1",
-    "mimmi20/json-class": "~2.0.8",
-    "psr/log": "^1.1.2",
-    "psr/simple-cache": "^1.0.1",
-    "symfony/finder": "^4.4.2",
-    "symfony/yaml": "^4.4.2"
+    "php": "^8.1.0",
+    "ext-mbstring": "*",
+    "psr/log": "^1.1.4 || ^3.0.0",
+    "psr/simple-cache": "^1.0.1 || ^3.0.0",
+    "symfony/finder": "^v6.0.2",
+    "symfony/yaml": "^v6.0.2"
   },
   "require-dev": {
     "ext-json": "*",
     "ext-pdo": "*",
     "ext-simplexml": "*",
-    "browscap/browscap": "dev-master",
-    "donatj/phpuseragentparser": "dev-master",
-    "ergebnis/composer-normalize": "^2.2.2",
-    "ergebnis/phpstan-rules": "^0.14.2",
-    "friendsofphp/php-cs-fixer": "^2.16.1",
-    "jaybizzle/crawler-detect": "dev-master",
-    "mimmi20/browser-detector": "~5.0.41",
-    "mobiledetect/mobiledetectlib": "dev-master",
-    "pepakriz/phpstan-exception-rules": "^0.10.1",
-    "phpstan/phpstan": "^0.12.5",
-    "phpstan/phpstan-deprecation-rules": "^0.12.2",
-    "phpstan/phpstan-strict-rules": "^0.12.1",
-    "piwik/device-detector": "dev-master",
-    "sinergi/browser-detector": "dev-master",
+    "browscap/browscap": "^6.0.x-dev",
+    "cbschuld/browser.php": "^v1.9.6",
+    "donatj/phpuseragentparser": "^v1.5.0",
+    "endorphin-studio/browser-detector": "^6.0.1",
+    "endorphin-studio/browser-detector-tests": "^2.2.0 || ^3.1.0",
+    "jaybizzle/crawler-detect": "^v1.2.110",
+    "matomo/device-detector": "dev-master",
+    "mimmi20/browser-detector": "^7.0.0",
+    "mimmi20/coding-standard": "^2.6.0",
+    "mobiledetect/mobiledetectlib": "^2.8.37",
+    "phpstan/extension-installer": "^1.1.0",
+    "phpstan/phpstan": "^1.3.1",
+    "phpstan/phpstan-deprecation-rules": "^1.0.0",
+    "sinergi/browser-detector": "^6.1.4",
     "squizlabs/php_codesniffer": "^3.5.3",
-    "ua-parser/uap-core": "dev-master",
-    "whichbrowser/parser": "dev-master",
-    "woothee/woothee-testset": "dev-master",
-    "yzalis/ua-parser": "dev-master",
-    "zsxsoft/php-useragent": "dev-master"
+    "whichbrowser/parser": "^v2.1.2",
+    "woothee/woothee-testset": "^v1.13.0",
+    "zsxsoft/php-useragent": "^1.5"
   },
   "suggest": {
     "ext-http": "to parse headers in the WhichBrowserSource",
