@@ -2,81 +2,80 @@
 /**
  * This file is part of the browscap-helper-source package.
  *
- * Copyright (c) 2016-2019, Thomas Mueller <mimmi20@live.de>
+ * Copyright (c) 2016-2022, Thomas Mueller <mimmi20@live.de>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
 declare(strict_types = 1);
+
 namespace BrowscapHelper\Source;
 
-use BrowscapHelper\Source\Ua\UserAgent;
-use Psr\Log\LoggerInterface;
+use PDO;
+use PDOStatement;
+use RuntimeException;
 
-final class PdoSource implements SourceInterface
+use function assert;
+use function is_array;
+use function trim;
+
+final class PdoSource implements OutputAwareInterface, SourceInterface
 {
+    use GetNameTrait;
     use GetUserAgentsTrait;
+    use OutputAwareTrait;
 
-    /**
-     * @var \PDO
-     */
-    private $pdo;
+    private const NAME = 'pdo-source';
 
-    /**
-     * @var \Psr\Log\LoggerInterface
-     */
-    private $logger;
+    private PDO $pdo;
 
-    /**
-     * @param \Psr\Log\LoggerInterface $logger
-     * @param \PDO                     $pdo
-     */
-    public function __construct(LoggerInterface $logger, \PDO $pdo)
+    public function __construct(PDO $pdo)
     {
-        $this->logger = $logger;
-        $this->pdo    = $pdo;
+        $this->pdo = $pdo;
     }
 
     /**
-     * @return string
+     * @throws void
+     *
+     * @phpcsSuppress SlevomatCodingStandard.Functions.UnusedParameter.UnusedParameter
      */
-    public function getName(): string
+    public function isReady(string $parentMessage): bool
     {
-        return 'PDO';
+        return true;
     }
 
     /**
-     * @return array[]|iterable
+     * @return iterable<array<mixed>>
+     * @phpstan-return iterable<array{headers: array<non-empty-string, non-empty-string>, device: array{deviceName: string|null, marketingName: string|null, manufacturer: string|null, brand: string|null, display: array{width: int|null, height: int|null, touch: bool|null, type: string|null, size: float|int|null}, type: string|null, ismobile: bool|null}, client: array{name: string|null, modus: string|null, version: string|null, manufacturer: string|null, bits: int|null, type: string|null, isbot: bool|null}, platform: array{name: string|null, marketingName: string|null, version: string|null, manufacturer: string|null, bits: int|null}, engine: array{name: string|null, version: string|null, manufacturer: string|null}}>
+     *
+     * @throws RuntimeException
+     *
+     * @phpcsSuppress SlevomatCodingStandard.Functions.UnusedParameter.UnusedParameter
      */
-    public function getHeaders(): iterable
+    public function getProperties(string $message, int &$messageLength = 0): iterable
     {
-        foreach ($this->getAgents() as $row) {
-            $ua    = UserAgent::fromUseragent(trim($row->agent));
-            $agent = (string) $ua;
+        $sql = 'SELECT DISTINCT SQL_BIG_RESULT HIGH_PRIORITY `agent` FROM `agents` ORDER BY `lastTimeFound` DESC, `count` DESC, `idAgents` DESC';
 
-            if (empty($agent)) {
+        $driverOptions = [PDO::ATTR_CURSOR => PDO::CURSOR_FWDONLY];
+
+        $stmt = $this->pdo->prepare($sql, $driverOptions);
+        assert($stmt instanceof PDOStatement);
+        $stmt->execute();
+
+        while ($row = $stmt->fetch(PDO::FETCH_ASSOC)) {
+            if (!is_array($row)) {
                 continue;
             }
 
-            yield $ua->getHeaders();
-        }
-    }
+            $agent = trim($row['agent']);
 
-    /**
-     * @return array[]|iterable
-     */
-    public function getProperties(): iterable
-    {
-        foreach ($this->getAgents() as $row) {
-            $ua    = UserAgent::fromUseragent(trim($row->agent));
-            $agent = (string) $ua;
-
-            if (empty($agent)) {
+            if ('' === $agent) {
                 continue;
             }
 
-            yield $agent => [
+            yield [
+                'headers' => ['user-agent' => $agent],
                 'device' => [
                     'deviceName' => null,
                     'marketingName' => null,
@@ -92,15 +91,9 @@ final class PdoSource implements SourceInterface
                     'dualOrientation' => null,
                     'type' => null,
                     'simCount' => null,
-                    'market' => [
-                        'regions' => null,
-                        'countries' => null,
-                        'vendors' => null,
-                    ],
-                    'connections' => null,
                     'ismobile' => null,
                 ],
-                'browser' => [
+                'client' => [
                     'name' => null,
                     'modus' => null,
                     'version' => null,
@@ -121,25 +114,9 @@ final class PdoSource implements SourceInterface
                     'version' => null,
                     'manufacturer' => null,
                 ],
+                'raw' => $row,
+                'file' => null,
             ];
-        }
-    }
-
-    /**
-     * @return iterable|\stdClass[]
-     */
-    private function getAgents(): iterable
-    {
-        $sql = 'SELECT DISTINCT SQL_BIG_RESULT HIGH_PRIORITY `agent` FROM `agents` ORDER BY `lastTimeFound` DESC, `count` DESC, `idAgents` DESC';
-
-        $driverOptions = [\PDO::ATTR_CURSOR => \PDO::CURSOR_FWDONLY];
-
-        /** @var \PDOStatement $stmt */
-        $stmt = $this->pdo->prepare($sql, $driverOptions);
-        $stmt->execute();
-
-        while ($row = $stmt->fetch(\PDO::FETCH_OBJ)) {
-            yield $row;
         }
     }
 }
