@@ -2,42 +2,45 @@
 /**
  * This file is part of the browscap-helper-source package.
  *
- * Copyright (c) 2016-2019, Thomas Mueller <mimmi20@live.de>
+ * Copyright (c) 2016-2022, Thomas Mueller <mimmi20@live.de>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
 declare(strict_types = 1);
+
 namespace BrowscapHelper\Source\Reader;
 
 use BrowscapHelper\Source\Helper\Regex;
-use Psr\Log\LoggerInterface;
+use BrowscapHelper\Source\OutputAwareInterface;
+use BrowscapHelper\Source\OutputAwareTrait;
+use Symfony\Component\Console\Output\OutputInterface;
 
-final class LogFileReader implements ReaderInterface
+use function array_key_exists;
+use function array_pop;
+use function explode;
+use function fclose;
+use function feof;
+use function fgets;
+use function fopen;
+use function mb_strlen;
+use function preg_match;
+use function sprintf;
+use function str_pad;
+use function trim;
+
+use const STR_PAD_RIGHT;
+
+final class LogFileReader implements OutputAwareInterface, ReaderInterface
 {
-    /**
-     * @var array
-     */
-    private $files = [];
+    use OutputAwareTrait;
+
+    /** @var array<string> */
+    private array $files = [];
 
     /**
-     * @var \Psr\Log\LoggerInterface
-     */
-    private $logger;
-
-    /**
-     * @param \Psr\Log\LoggerInterface $logger
-     */
-    public function __construct(LoggerInterface $logger)
-    {
-        $this->logger = $logger;
-    }
-
-    /**
-     * @param string $file
-     *
-     * @return void
+     * @throws void
      */
     public function addLocalFile(string $file): void
     {
@@ -45,21 +48,28 @@ final class LogFileReader implements ReaderInterface
     }
 
     /**
-     * @param \Psr\Log\LoggerInterface $logger
+     * @return array<string>|iterable
      *
-     * @return \Generator
+     * @throws void
      */
-    public function getAgents(LoggerInterface $logger): iterable
+    public function getAgents(string $parentMessage = '', int &$messageLength = 0): iterable
     {
         $regex = (new Regex())->getRegex();
 
         foreach ($this->files as $file) {
-            $this->logger->info('    reading file ' . str_pad($file, 100, ' ', STR_PAD_RIGHT));
+            $message = $parentMessage . sprintf('- reading file %s', $file);
+
+            if (mb_strlen($message) > $messageLength) {
+                $messageLength = mb_strlen($message);
+            }
+
+            $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERBOSE);
 
             $handle = @fopen($file, 'r');
 
             if (false === $handle) {
-                $this->logger->emergency(new \RuntimeException('reading file ' . $file . ' caused an error'));
+                $this->writeln('', OutputInterface::VERBOSITY_VERBOSE);
+                $this->writeln("\r" . '<error>' . $parentMessage . sprintf('- reading file %s caused an error</error>', $file), OutputInterface::VERBOSITY_NORMAL);
                 continue;
             }
 
@@ -71,6 +81,7 @@ final class LogFileReader implements ReaderInterface
                 if (false === $line) {
                     continue;
                 }
+
                 ++$i;
 
                 if (empty($line)) {
@@ -80,7 +91,7 @@ final class LogFileReader implements ReaderInterface
                 $lineMatches = [];
 
                 if (!(bool) preg_match($regex, $line, $lineMatches)) {
-                    $logger->error('no useragent found in line "' . $line . '" used regex: "' . $regex . '"');
+                    $this->writeln("\r" . '<error>' . $parentMessage . sprintf('- no useragent found in line "%s" used regex: "%s"</error>', $line, $regex), OutputInterface::VERBOSITY_NORMAL);
 
                     continue;
                 }
@@ -95,7 +106,7 @@ final class LogFileReader implements ReaderInterface
                     continue;
                 }
 
-                yield $agentOfLine;
+                yield $file => $agentOfLine;
             }
 
             fclose($handle);
@@ -103,9 +114,7 @@ final class LogFileReader implements ReaderInterface
     }
 
     /**
-     * @param string $text
-     *
-     * @return string
+     * @throws void
      */
     private function extractAgent(string $text): string
     {
