@@ -14,34 +14,38 @@ namespace BrowscapHelper\Source;
 
 use FilterIterator;
 use Iterator;
+use JsonException;
+use LogicException;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
 use RuntimeException;
 use SplFileInfo;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Yaml\Yaml;
+use UaDeviceType\TypeLoader;
 
-use function array_key_exists;
 use function assert;
 use function file_exists;
+use function file_get_contents;
 use function is_array;
 use function is_string;
+use function json_decode;
 use function mb_strlen;
 use function sprintf;
 use function str_pad;
 use function str_replace;
-use function trim;
 
+use const JSON_THROW_ON_ERROR;
+use const PHP_EOL;
 use const STR_PAD_RIGHT;
 
-final class WootheeSource implements OutputAwareInterface, SourceInterface
+final class BrowserDetectorSource implements OutputAwareInterface, SourceInterface
 {
     use GetNameTrait;
     use GetUserAgentsTrait;
     use OutputAwareTrait;
 
-    private const NAME = 'woothee/woothee-testset';
-    private const PATH = 'vendor/woothee/woothee-testset/testsets';
+    private const NAME = 'mimmi20/browser-detector';
+    private const PATH = 'vendor/mimmi20/browser-detector/tests/data';
 
     /**
      * @throws void
@@ -61,6 +65,7 @@ final class WootheeSource implements OutputAwareInterface, SourceInterface
      * @return iterable<array<mixed>>
      * @phpstan-return iterable<array{headers: array<non-empty-string, non-empty-string>, device: array{deviceName: string|null, marketingName: string|null, manufacturer: string|null, brand: string|null, display: array{width: int|null, height: int|null, touch: bool|null, type: string|null, size: float|int|null}, type: string|null, ismobile: bool|null}, client: array{name: string|null, modus: string|null, version: string|null, manufacturer: string|null, bits: int|null, type: string|null, isbot: bool|null}, platform: array{name: string|null, marketingName: string|null, version: string|null, manufacturer: string|null, bits: int|null}, engine: array{name: string|null, version: string|null, manufacturer: string|null}}>
      *
+     * @throws LogicException
      * @throws RuntimeException
      */
     public function getProperties(string $parentMessage, int &$messageLength = 0): iterable
@@ -74,7 +79,7 @@ final class WootheeSource implements OutputAwareInterface, SourceInterface
         $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERBOSE);
 
         $iterator = new RecursiveIteratorIterator(new RecursiveDirectoryIterator(self::PATH));
-        $files    = new class ($iterator, 'yaml') extends FilterIterator {
+        $files    = new class ($iterator, 'json') extends FilterIterator {
             private string $extension;
 
             /**
@@ -110,64 +115,75 @@ final class WootheeSource implements OutputAwareInterface, SourceInterface
 
             $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERY_VERBOSE);
 
-            $data = Yaml::parseFile($filepath);
+            $content = file_get_contents($filepath);
+
+            if (false === $content || '' === $content || PHP_EOL === $content) {
+                continue;
+            }
+
+            try {
+                $data = json_decode($content, true, 512, JSON_THROW_ON_ERROR);
+            } catch (JsonException $e) {
+                $this->writeln('', OutputInterface::VERBOSITY_VERBOSE);
+                $this->writeln('    <error>parsing file content [' . $filepath . '] failed</error>', OutputInterface::VERBOSITY_NORMAL);
+
+                continue;
+            }
 
             if (!is_array($data)) {
                 continue;
             }
 
-            foreach ($data as $row) {
-                if (!array_key_exists('target', $row) || empty($row['target'])) {
+            foreach ($data as $test) {
+                if (!is_array($test['headers']) || !isset($test['headers']['user-agent'])) {
                     continue;
                 }
 
-                $agent = trim($row['target']);
-
-                if ('' === $agent) {
+                if ('this is a fake ua to trigger the fallback' === $test['headers']['user-agent']) {
                     continue;
                 }
 
                 yield [
-                    'headers' => ['user-agent' => $agent],
+                    'headers' => $test['headers'],
                     'device' => [
-                        'deviceName' => null,
-                        'marketingName' => null,
-                        'manufacturer' => null,
-                        'brand' => null,
+                        'deviceName' => $test['result']['device']['deviceName'],
+                        'marketingName' => $test['result']['device']['marketingName'],
+                        'manufacturer' => $test['result']['device']['manufacturer'],
+                        'brand' => $test['result']['device']['brand'],
                         'display' => [
-                            'width' => null,
-                            'height' => null,
-                            'touch' => null,
-                            'type' => null,
-                            'size' => null,
+                            'width' => $test['result']['device']['display']['width'],
+                            'height' => $test['result']['device']['display']['height'],
+                            'touch' => $test['result']['device']['display']['touch'],
+                            'type' => $test['result']['device']['display']['type'] ?? null,
+                            'size' => $test['result']['device']['display']['size'],
                         ],
-                        'dualOrientation' => null,
-                        'type' => !isset($row['category']) || 'crawler' === $row['category'] ? null : $row['category'],
-                        'simCount' => null,
-                        'ismobile' => null,
+                        'dualOrientation' => $test['result']['device']['dualOrientation'] ?? null,
+                        'type' => $test['result']['device']['type'],
+                        'simCount' => $test['result']['device']['simCount'] ?? null,
+                        'ismobile' => (new TypeLoader())->load($test['result']['device']['type'])->isMobile(),
                     ],
                     'client' => [
-                        'name' => $row['name'] ?? null,
-                        'modus' => null,
-                        'version' => $row['version'] ?? null,
-                        'manufacturer' => null,
-                        'bits' => null,
-                        'type' => $row['category'] ?? null,
-                        'isbot' => isset($row['category']) && 'crawler' === $row['category'],
+                        'name' => $test['result']['browser']['name'],
+                        'modus' => $test['result']['browser']['modus'],
+                        'version' => ('0.0.0' === $test['result']['browser']['version'] ? null : $test['result']['browser']['version']),
+                        'manufacturer' => $test['result']['browser']['manufacturer'],
+                        'bits' => $test['result']['browser']['bits'],
+                        'type' => $test['result']['browser']['type'],
+                        'isbot' => (new \UaBrowserType\TypeLoader())->load($test['result']['browser']['type'])->isBot(),
                     ],
                     'platform' => [
-                        'name' => isset($row['os']) && 'UNKNOWN' !== $row['os'] ? $row['os'] : null,
-                        'marketingName' => null,
-                        'version' => $row['os_version'] ?? null,
-                        'manufacturer' => null,
-                        'bits' => null,
+                        'name' => $test['result']['os']['name'],
+                        'marketingName' => $test['result']['os']['marketingName'],
+                        'version' => ('0.0.0' === $test['result']['os']['version'] ? null : $test['result']['os']['version']),
+                        'manufacturer' => $test['result']['os']['manufacturer'],
+                        'bits' => $test['result']['os']['bits'],
                     ],
                     'engine' => [
-                        'name' => null,
-                        'version' => null,
-                        'manufacturer' => null,
+                        'name' => $test['result']['engine']['name'],
+                        'version' => $test['result']['engine']['version'],
+                        'manufacturer' => $test['result']['engine']['manufacturer'],
                     ],
-                    'raw' => $row,
+                    'raw' => $test['result'],
                     'file' => $filepath,
                 ];
             }
