@@ -12,28 +12,35 @@ declare(strict_types = 1);
 
 namespace BrowscapHelper\Source;
 
-use LogicException;
+use FilterIterator;
+use Iterator;
+use RecursiveDirectoryIterator;
+use RecursiveIteratorIterator;
 use RuntimeException;
+use SplFileInfo;
 use Symfony\Component\Console\Output\OutputInterface;
+use UnexpectedValueException;
 
+use function assert;
 use function file;
 use function file_exists;
+use function is_string;
 use function mb_strlen;
 use function sprintf;
+use function str_getcsv;
 use function str_pad;
+use function str_replace;
 
-use const FILE_IGNORE_NEW_LINES;
-use const FILE_SKIP_EMPTY_LINES;
 use const STR_PAD_RIGHT;
 
-final class CrawlerDetectSource implements OutputAwareInterface, SourceInterface
+final class CbschuldSource implements OutputAwareInterface, SourceInterface
 {
     use GetNameTrait;
     use GetUserAgentsTrait;
     use OutputAwareTrait;
 
-    private const NAME = 'jaybizzle/crawler-detect';
-    private const PATH = 'vendor/jaybizzle/crawler-detect/tests';
+    private const NAME = 'cbschuld/browser.php';
+    private const PATH = 'vendor/cbschuld/browser.php/tests/lists';
 
     /**
      * @throws void
@@ -53,8 +60,8 @@ final class CrawlerDetectSource implements OutputAwareInterface, SourceInterface
      * @return iterable<array<mixed>>
      * @phpstan-return iterable<array{headers: array<non-empty-string, non-empty-string>, device: array{deviceName: string|null, marketingName: string|null, manufacturer: string|null, brand: string|null, display: array{width: int|null, height: int|null, touch: bool|null, type: string|null, size: float|int|null}, type: string|null, ismobile: bool|null}, client: array{name: string|null, modus: string|null, version: string|null, manufacturer: string|null, bits: int|null, type: string|null, isbot: bool|null}, platform: array{name: string|null, marketingName: string|null, version: string|null, manufacturer: string|null, bits: int|null}, engine: array{name: string|null, version: string|null, manufacturer: string|null}}>
      *
-     * @throws LogicException
      * @throws RuntimeException
+     * @throws UnexpectedValueException
      */
     public function getProperties(string $parentMessage, int &$messageLength = 0): iterable
     {
@@ -66,26 +73,58 @@ final class CrawlerDetectSource implements OutputAwareInterface, SourceInterface
 
         $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERBOSE);
 
-        $filepath = self::PATH . '/crawlers.txt';
+        $iterator = new RecursiveIteratorIterator(new RecursiveDirectoryIterator(self::PATH));
+        $files    = new class ($iterator, 'txt') extends FilterIterator {
+            private string $extension;
 
-        $lines = file($filepath, FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
+            /**
+             * @param Iterator<SplFileInfo> $iterator
+             */
+            public function __construct(Iterator $iterator, string $extension)
+            {
+                parent::__construct($iterator);
+                $this->extension = $extension;
+            }
 
-        $message = $parentMessage . sprintf('- reading file %s', $filepath);
+            public function accept(): bool
+            {
+                $file = $this->getInnerIterator()->current();
 
-        if (mb_strlen($message) > $messageLength) {
-            $messageLength = mb_strlen($message);
-        }
+                assert($file instanceof SplFileInfo);
 
-        $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERY_VERBOSE);
+                return $file->isFile() && $file->getExtension() === $this->extension;
+            }
+        };
 
-        if (false !== $lines) {
-            foreach ($lines as $ua) {
-                if (empty($ua)) {
+        foreach ($files as $file) {
+            /** @var SplFileInfo $file */
+            $pathName = $file->getPathname();
+            $filepath = str_replace('\\', '/', $pathName);
+            assert(is_string($filepath));
+
+            $message = $parentMessage . sprintf('- reading file %s', $filepath);
+
+            if (mb_strlen($message) > $messageLength) {
+                $messageLength = mb_strlen($message);
+            }
+
+            $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERY_VERBOSE);
+
+            $lines = file($filepath);
+
+            if (false === $lines) {
+                continue;
+            }
+
+            foreach ($lines as $line) {
+                $testData = str_getcsv($line, "\t");
+
+                if (empty($testData[0])) {
                     continue;
                 }
 
                 yield [
-                    'headers' => ['user-agent' => $ua],
+                    'headers' => ['user-agent' => $testData[0]],
                     'device' => [
                         'deviceName' => null,
                         'marketingName' => null,
@@ -104,16 +143,16 @@ final class CrawlerDetectSource implements OutputAwareInterface, SourceInterface
                         'ismobile' => null,
                     ],
                     'client' => [
-                        'name' => null,
+                        'name' => $testData[2],
                         'modus' => null,
-                        'version' => null,
+                        'version' => $testData[3],
                         'manufacturer' => null,
                         'bits' => null,
                         'type' => null,
-                        'isbot' => true,
+                        'isbot' => null,
                     ],
                     'platform' => [
-                        'name' => null,
+                        'name' => $testData[5] ?? null,
                         'marketingName' => null,
                         'version' => null,
                         'manufacturer' => null,
@@ -124,76 +163,10 @@ final class CrawlerDetectSource implements OutputAwareInterface, SourceInterface
                         'version' => null,
                         'manufacturer' => null,
                     ],
-                    'raw' => $ua,
+                    'raw' => $testData,
                     'file' => $filepath,
                 ];
             }
-        }
-
-        $filepath = self::PATH . '/devices.txt';
-
-        $lines = file($filepath, FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
-
-        $message = $parentMessage . sprintf('- reading file %s', $filepath);
-
-        if (mb_strlen($message) > $messageLength) {
-            $messageLength = mb_strlen($message);
-        }
-
-        $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERY_VERBOSE);
-
-        if (false === $lines) {
-            return;
-        }
-
-        foreach ($lines as $ua) {
-            if (empty($ua)) {
-                continue;
-            }
-
-            yield [
-                'headers' => ['user-agent' => $ua],
-                'device' => [
-                    'deviceName' => null,
-                    'marketingName' => null,
-                    'manufacturer' => null,
-                    'brand' => null,
-                    'display' => [
-                        'width' => null,
-                        'height' => null,
-                        'touch' => null,
-                        'type' => null,
-                        'size' => null,
-                    ],
-                    'dualOrientation' => null,
-                    'type' => null,
-                    'simCount' => null,
-                    'ismobile' => null,
-                ],
-                'client' => [
-                    'name' => null,
-                    'modus' => null,
-                    'version' => null,
-                    'manufacturer' => null,
-                    'bits' => null,
-                    'type' => null,
-                    'isbot' => false,
-                ],
-                'platform' => [
-                    'name' => null,
-                    'marketingName' => null,
-                    'version' => null,
-                    'manufacturer' => null,
-                    'bits' => null,
-                ],
-                'engine' => [
-                    'name' => null,
-                    'version' => null,
-                    'manufacturer' => null,
-                ],
-                'raw' => $ua,
-                'file' => $filepath,
-            ];
         }
     }
 }
