@@ -12,8 +12,11 @@ declare(strict_types = 1);
 
 namespace BrowscapHelper\Source;
 
+use DeviceDetector\Parser\Client\Browser;
+use DeviceDetector\Parser\Device\AbstractDeviceParser;
 use FilterIterator;
 use Iterator;
+use LogicException;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
 use RuntimeException;
@@ -21,9 +24,12 @@ use SplFileInfo;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Yaml\Yaml;
 
-use function array_key_exists;
+use function array_map;
 use function assert;
+use function explode;
 use function file_exists;
+use function implode;
+use function in_array;
 use function is_array;
 use function is_string;
 use function mb_strlen;
@@ -34,14 +40,14 @@ use function trim;
 
 use const STR_PAD_RIGHT;
 
-final class WootheeSource implements OutputAwareInterface, SourceInterface
+final class MatomoSource implements OutputAwareInterface, SourceInterface
 {
     use GetNameTrait;
     use GetUserAgentsTrait;
     use OutputAwareTrait;
 
-    private const NAME = 'woothee/woothee-testset';
-    private const PATH = 'vendor/woothee/woothee-testset/testsets';
+    private const NAME = 'matomo/device-detector';
+    private const PATH = 'vendor/matomo/device-detector/regexes';
 
     /**
      * @throws void
@@ -61,6 +67,7 @@ final class WootheeSource implements OutputAwareInterface, SourceInterface
      * @return iterable<array<mixed>>
      * @phpstan-return iterable<array{headers: array<non-empty-string, non-empty-string>, device: array{deviceName: string|null, marketingName: string|null, manufacturer: string|null, brand: string|null, display: array{width: int|null, height: int|null, touch: bool|null, type: string|null, size: float|int|null}, type: string|null, ismobile: bool|null}, client: array{name: string|null, modus: string|null, version: string|null, manufacturer: string|null, bits: int|null, type: string|null, isbot: bool|null}, platform: array{name: string|null, marketingName: string|null, version: string|null, manufacturer: string|null, bits: int|null}, engine: array{name: string|null, version: string|null, manufacturer: string|null}}>
      *
+     * @throws LogicException
      * @throws RuntimeException
      */
     public function getProperties(string $parentMessage, int &$messageLength = 0): iterable
@@ -74,7 +81,7 @@ final class WootheeSource implements OutputAwareInterface, SourceInterface
         $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERBOSE);
 
         $iterator = new RecursiveIteratorIterator(new RecursiveDirectoryIterator(self::PATH));
-        $files    = new class ($iterator, 'yaml') extends FilterIterator {
+        $files    = new class ($iterator, 'yml') extends FilterIterator {
             private string $extension;
 
             /**
@@ -117,23 +124,25 @@ final class WootheeSource implements OutputAwareInterface, SourceInterface
             }
 
             foreach ($data as $row) {
-                if (!array_key_exists('target', $row) || empty($row['target'])) {
+                if (empty($row['user_agent'])) {
                     continue;
                 }
 
-                $agent = trim($row['target']);
+                $ua    = explode("\n", $row['user_agent']);
+                $ua    = array_map('trim', $ua);
+                $agent = trim(implode(' ', $ua));
 
-                if ('' === $agent) {
+                if (empty($agent)) {
                     continue;
                 }
 
                 yield [
                     'headers' => ['user-agent' => $agent],
                     'device' => [
-                        'deviceName' => null,
+                        'deviceName' => $row['device']['model'] ?? null,
                         'marketingName' => null,
                         'manufacturer' => null,
-                        'brand' => null,
+                        'brand' => (!empty($row['device']['brand']) ? AbstractDeviceParser::getFullName($row['device']['brand']) : null),
                         'display' => [
                             'width' => null,
                             'height' => null,
@@ -142,29 +151,29 @@ final class WootheeSource implements OutputAwareInterface, SourceInterface
                             'size' => null,
                         ],
                         'dualOrientation' => null,
-                        'type' => !isset($row['category']) || 'crawler' === $row['category'] ? null : $row['category'],
+                        'type' => $row['device']['type'] ?? null,
                         'simCount' => null,
-                        'ismobile' => null,
+                        'ismobile' => $this->isMobile($row),
                     ],
                     'client' => [
-                        'name' => $row['name'] ?? null,
+                        'name' => $data['bot'] ? ($data['bot']['name'] ?? null) : ($row['client']['name'] ?? null),
                         'modus' => null,
-                        'version' => $row['version'] ?? null,
+                        'version' => $data['bot'] ? null : ($data['client']['version'] ?? null),
                         'manufacturer' => null,
                         'bits' => null,
-                        'type' => $row['category'] ?? null,
-                        'isbot' => isset($row['category']) && 'crawler' === $row['category'],
+                        'type' => $data['bot'] ? ($data['bot']['category'] ?? null) : ($data['client']['type'] ?? null),
+                        'isbot' => !empty($data['bot']),
                     ],
                     'platform' => [
-                        'name' => isset($row['os']) && 'UNKNOWN' !== $row['os'] ? $row['os'] : null,
+                        'name' => $row['os']['name'] ?? null,
                         'marketingName' => null,
-                        'version' => $row['os_version'] ?? null,
+                        'version' => $row['os']['version'] ?? null,
                         'manufacturer' => null,
                         'bits' => null,
                     ],
                     'engine' => [
-                        'name' => null,
-                        'version' => null,
+                        'name' => (!empty($row['client']['engine']) ? $row['client']['engine'] : null),
+                        'version' => (!empty($row['client']['engine_version']) ? $row['client']['engine_version'] : null),
                         'manufacturer' => null,
                     ],
                     'raw' => $row,
@@ -172,5 +181,87 @@ final class WootheeSource implements OutputAwareInterface, SourceInterface
                 ];
             }
         }
+    }
+
+    /**
+     * @param array<mixed> $data
+     * @phpstan-param array{os: array{short_name: string|null}, client: array{type: string, short_name?: string}, os_family: string, device: array{type?: int}} $data
+     */
+    private function isMobile(array $data): bool
+    {
+        if (empty($data['device']['type'])) {
+            return false;
+        }
+
+        $device     = $data['device']['type'];
+        $deviceType = AbstractDeviceParser::getAvailableDeviceTypes()[$device];
+
+        if (!empty($deviceType)) {
+            // Mobile device types
+            if (
+                in_array(
+                    $deviceType,
+                    [
+                        AbstractDeviceParser::DEVICE_TYPE_FEATURE_PHONE,
+                        AbstractDeviceParser::DEVICE_TYPE_SMARTPHONE,
+                        AbstractDeviceParser::DEVICE_TYPE_TABLET,
+                        AbstractDeviceParser::DEVICE_TYPE_PHABLET,
+                        AbstractDeviceParser::DEVICE_TYPE_CAMERA,
+                        AbstractDeviceParser::DEVICE_TYPE_PORTABLE_MEDIA_PAYER,
+                    ],
+                    true
+                )
+            ) {
+                return true;
+            }
+
+            // non mobile device types
+            if (
+                in_array(
+                    $deviceType,
+                    [
+                        AbstractDeviceParser::DEVICE_TYPE_TV,
+                        AbstractDeviceParser::DEVICE_TYPE_SMART_DISPLAY,
+                        AbstractDeviceParser::DEVICE_TYPE_CONSOLE,
+                    ],
+                    true
+                )
+            ) {
+                return false;
+            }
+        }
+
+        // Check for browsers available for mobile devices only
+        if ('browser' === $data['client']['type'] && Browser::isMobileOnlyBrowser($data['client']['short_name'] ?? 'UNK')) {
+            return true;
+        }
+
+        $osShort = $data['os']['short_name'];
+
+        if (empty($osShort) || 'UNK' === $osShort) {
+            return false;
+        }
+
+        return !$this->isDesktop($data);
+    }
+
+    /**
+     * @param array<mixed> $data
+     * @phpstan-param array{os: array{short_name: string|null}, client: array{type: string, short_name?: string}, os_family: string, device: array{type?: int}} $data
+     */
+    private function isDesktop(array $data): bool
+    {
+        $osShort = $data['os']['short_name'];
+
+        if (empty($osShort) || 'UNK' === $osShort) {
+            return false;
+        }
+
+        // Check for browsers available for mobile devices only
+        if ('browser' === $data['client']['type'] && Browser::isMobileOnlyBrowser($data['client']['short_name'] ?? 'UNK')) {
+            return false;
+        }
+
+        return in_array($data['os_family'], ['AmigaOS', 'IBM', 'GNU/Linux', 'Mac', 'Unix', 'Windows', 'BeOS', 'Chrome OS'], true);
     }
 }
