@@ -2,14 +2,18 @@
 /**
  * This file is part of the browscap-helper-source package.
  *
- * Copyright (c) 2016-2019, Thomas Mueller <mimmi20@live.de>
+ * Copyright (c) 2016-2022, Thomas Mueller <mimmi20@live.de>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
 declare(strict_types = 1);
+
 namespace BrowscapHelper\Source;
+
+use LogicException;
+use RuntimeException;
 
 interface SourceInterface
 {
@@ -17,32 +21,35 @@ interface SourceInterface
 
     public const DELIMETER_HEADER_ROW = '::==::';
 
-    /**
-     * @throws \LogicException
-     * @throws \RuntimeException
-     *
-     * @return iterable|string[]
-     */
-    public function getUserAgents(): iterable;
+    public function isReady(string $parentMessage): bool;
 
     /**
-     * @throws \LogicException
-     * @throws \RuntimeException
+     * @return iterable<int, non-empty-string>
      *
-     * @return array[]|iterable
+     * @throws LogicException
+     * @throws RuntimeException
      */
-    public function getHeaders(): iterable;
+    public function getUserAgents(string $message, int &$messageLength = 0): iterable;
 
     /**
-     * @return string
+     * @return iterable<array<non-empty-string, non-empty-string>>
+     *
+     * @throws LogicException
+     * @throws RuntimeException
+     */
+    public function getHeaders(string $message, int &$messageLength = 0): iterable;
+
+    /**
+     * @return iterable<array<mixed>>
+     * @phpstan-return iterable<array{headers: array<non-empty-string, non-empty-string>, device: array{deviceName: string|null, marketingName: string|null, manufacturer: string|null, brand: string|null, display: array{width: int|null, height: int|null, touch: bool|null, type: string|null, size: float|int|null}, type: string|null, ismobile: bool|null}, client: array{name: string|null, modus: string|null, version: string|null, manufacturer: string|null, bits: int|null, type: string|null, isbot: bool|null}, platform: array{name: string|null, marketingName: string|null, version: string|null, manufacturer: string|null, bits: int|null}, engine: array{name: string|null, version: string|null, manufacturer: string|null}}>
+     *
+     * @throws LogicException
+     * @throws RuntimeException
+     */
+    public function getProperties(string $parentMessage, int &$messageLength = 0): iterable;
+
+    /**
+     * @throws void
      */
     public function getName(): string;
-
-    /**
-     * @throws \LogicException
-     * @throws \RuntimeException
-     *
-     * @return array[]|iterable
-     */
-    public function getProperties(): iterable;
 }
