@@ -2,86 +2,132 @@
 /**
  * This file is part of the browscap-helper-source package.
  *
- * Copyright (c) 2016-2019, Thomas Mueller <mimmi20@live.de>
+ * Copyright (c) 2016-2022, Thomas Mueller <mimmi20@live.de>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
 declare(strict_types = 1);
+
 namespace BrowscapHelper\Source;
 
-use BrowscapHelper\Source\Ua\UserAgent;
-use Psr\Log\LoggerInterface;
-use Symfony\Component\Finder\Finder;
+use LogicException;
+use RuntimeException;
+use SplFileObject;
+use Symfony\Component\Console\Output\OutputInterface;
 
-final class ZsxsoftSource implements SourceInterface
+use function array_unique;
+use function count;
+use function file_exists;
+use function mb_strlen;
+use function mb_strpos;
+use function preg_match;
+use function sprintf;
+use function str_pad;
+use function str_replace;
+use function trim;
+use function usort;
+
+use const STR_PAD_RIGHT;
+
+final class ZsxsoftSource implements OutputAwareInterface, SourceInterface
 {
+    use GetNameTrait;
     use GetUserAgentsTrait;
+    use OutputAwareTrait;
+
+    private const NAME = 'zsxsoft/php-useragent';
+    private const PATH = 'vendor/zsxsoft/php-useragent/lib';
 
     /**
-     * @var \Psr\Log\LoggerInterface
+     * @throws void
      */
-    private $logger;
-
-    /**
-     * @param \Psr\Log\LoggerInterface $logger
-     */
-    public function __construct(LoggerInterface $logger)
+    public function isReady(string $parentMessage): bool
     {
-        $this->logger = $logger;
-    }
-
-    /**
-     * @return string
-     */
-    public function getName(): string
-    {
-        return 'zsxsoft/php-useragent';
-    }
-
-    /**
-     * @throws \LogicException
-     *
-     * @return array[]|iterable
-     */
-    public function getHeaders(): iterable
-    {
-        foreach ($this->loadFromPath() as $data) {
-            $agent = trim($data[0][0]);
-
-            $ua    = UserAgent::fromUseragent($agent);
-            $agent = (string) $ua;
-
-            if (empty($agent)) {
-                continue;
-            }
-
-            yield $ua->getHeaders();
+        if (file_exists(self::PATH)) {
+            return true;
         }
+
+        $this->writeln("\r" . '<error>' . $parentMessage . sprintf('- path %s not found</error>', self::PATH), OutputInterface::VERBOSITY_NORMAL);
+
+        return false;
     }
 
     /**
-     * @throws \LogicException
-     * @throws \RuntimeException
+     * @return iterable<array<mixed>>
+     * @phpstan-return iterable<array{headers: array<non-empty-string, non-empty-string>, device: array{deviceName: string|null, marketingName: string|null, manufacturer: string|null, brand: string|null, display: array{width: int|null, height: int|null, touch: bool|null, type: string|null, size: float|int|null}, type: string|null, ismobile: bool|null}, client: array{name: string|null, modus: string|null, version: string|null, manufacturer: string|null, bits: int|null, type: string|null, isbot: bool|null}, platform: array{name: string|null, marketingName: string|null, version: string|null, manufacturer: string|null, bits: int|null}, engine: array{name: string|null, version: string|null, manufacturer: string|null}}>
      *
-     * @return array[]|iterable
+     * @throws LogicException
+     * @throws RuntimeException
      */
-    public function getProperties(): iterable
+    public function getProperties(string $parentMessage, int &$messageLength = 0): iterable
     {
         $brands = $this->getBrands();
 
-        foreach ($this->loadFromPath() as $data) {
-            $agent = trim($data[0][0]);
+        $message = $parentMessage . sprintf('- reading path %s', self::PATH);
 
-            $ua    = UserAgent::fromUseragent($agent);
-            $agent = (string) $ua;
+        if (mb_strlen($message) > $messageLength) {
+            $messageLength = mb_strlen($message);
+        }
 
-            if (empty($agent)) {
+        $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERBOSE);
+
+        $brands = [];
+        $file   = new SplFileObject('vendor/zsxsoft/php-useragent/lib/useragent_detect_device.php');
+        $file->setFlags(SplFileObject::DROP_NEW_LINE);
+        while (!$file->eof()) {
+            $line = $file->fgets();
+
+            if (false === $line) {
                 continue;
             }
 
-            $model = '';
+            $line = trim($line);
+            preg_match('/^\$brand = ("|\')(.*)("|\');$/', $line, $matches);
+
+            if (0 >= count($matches)) {
+                continue;
+            }
+
+            $brand = $matches[2];
+
+            if (empty($brand)) {
+                continue;
+            }
+
+            $brands[] = $brand;
+        }
+
+        $brands = array_unique($brands);
+
+        usort($brands, static fn ($a, $b) => mb_strlen($b) - mb_strlen($a));
+
+        $filepath = 'vendor/zsxsoft/php-useragent/tests/UserAgentList.php';
+
+        $message = $parentMessage . sprintf('- reading file %s', $filepath);
+
+        if (mb_strlen($message) > $messageLength) {
+            $messageLength = mb_strlen($message);
+        }
+
+        $this->write("\r" . '<info>' . str_pad($message, $messageLength, ' ', STR_PAD_RIGHT) . '</info>', false, OutputInterface::VERBOSITY_VERY_VERBOSE);
+
+        $provider = include $filepath;
+
+        foreach ($provider as $data) {
+            if (!isset($data[0][0])) {
+                continue;
+            }
+
+            $agent = trim($data[0][0]);
+
+            if ('' === $agent) {
+                continue;
+            }
+
+            $model = null;
+            $brand = null;
 
             foreach ($brands as $brand) {
                 if (false !== mb_strpos($data[1][8], $brand)) {
@@ -90,15 +136,16 @@ final class ZsxsoftSource implements SourceInterface
                     break;
                 }
 
-                $brand = '';
+                $brand = null;
             }
 
-            yield $agent => [
+            yield [
+                'headers' => ['user-agent' => $agent],
                 'device' => [
-                    'deviceName' => $model,
+                    'deviceName' => empty($model) ? null : $model,
                     'marketingName' => null,
                     'manufacturer' => null,
-                    'brand' => $brand ?? '',
+                    'brand' => empty($brand) ? null : $brand,
                     'display' => [
                         'width' => null,
                         'height' => null,
@@ -109,27 +156,21 @@ final class ZsxsoftSource implements SourceInterface
                     'dualOrientation' => null,
                     'type' => null,
                     'simCount' => null,
-                    'market' => [
-                        'regions' => null,
-                        'countries' => null,
-                        'vendors' => null,
-                    ],
-                    'connections' => null,
                     'ismobile' => null,
                 ],
-                'browser' => [
-                    'name' => $data[1][2],
+                'client' => [
+                    'name' => empty($data[1][2]) ? null : $data[1][2],
                     'modus' => null,
-                    'version' => $data[1][3],
+                    'version' => empty($data[1][3]) ? null : $data[1][3],
                     'manufacturer' => null,
                     'bits' => null,
                     'type' => null,
                     'isbot' => null,
                 ],
                 'platform' => [
-                    'name' => $data[1][5],
+                    'name' => empty($data[1][5]) ? null : $data[1][5],
                     'marketingName' => null,
-                    'version' => $data[1][6],
+                    'version' => empty($data[1][6]) ? null : $data[1][6],
                     'manufacturer' => null,
                     'bits' => null,
                 ],
@@ -138,77 +179,47 @@ final class ZsxsoftSource implements SourceInterface
                     'version' => null,
                     'manufacturer' => null,
                 ],
+                'raw' => $data,
+                'file' => $filepath,
             ];
         }
     }
 
     /**
-     * @throws \LogicException
-     *
-     * @return array[]|iterable
-     */
-    private function loadFromPath(): iterable
-    {
-        $path = 'vendor/zsxsoft/php-useragent/tests';
-
-        if (!file_exists($path)) {
-            $this->logger->warning(sprintf('    path %s not found', $path));
-
-            return;
-        }
-
-        $this->logger->info(sprintf('    reading path %s', $path));
-
-        $finder = new Finder();
-        $finder->files();
-        $finder->name('UserAgentList.php');
-        $finder->ignoreDotFiles(true);
-        $finder->ignoreVCS(true);
-        $finder->sortByName();
-        $finder->ignoreUnreadableDirs();
-        $finder->in($path);
-
-        foreach ($finder as $file) {
-            /** @var \Symfony\Component\Finder\SplFileInfo $file */
-            $filepath = $file->getPathname();
-
-            $this->logger->info('    reading file ' . str_pad($filepath, 100, ' ', STR_PAD_RIGHT));
-
-            $provider = require $filepath;
-
-            foreach ($provider as $data) {
-                yield $data;
-            }
-        }
-    }
-
-    /**
-     * @throws \LogicException
-     * @throws \RuntimeException
-     *
      * @return string[]
+     *
+     * @throws LogicException
+     * @throws RuntimeException
      */
     private function getBrands(): array
     {
         $brands = [];
-        $file   = new \SplFileObject('vendor/zsxsoft/php-useragent/lib/useragent_detect_device.php');
-        $file->setFlags(\SplFileObject::DROP_NEW_LINE);
+        $file   = new SplFileObject('vendor/zsxsoft/php-useragent/lib/useragent_detect_device.php');
+        $file->setFlags(SplFileObject::DROP_NEW_LINE);
         while (!$file->eof()) {
-            $line = trim($file->fgets());
+            $line = $file->fgets();
+            if (false === $line) {
+                continue;
+            }
+
+            $line = trim($line);
             preg_match('/^\$brand = (["\'])(.*)(["\']);$/', $line, $matches);
 
-            if (0 < count($matches)) {
-                $brand = $matches[2];
-                if (!empty($brand)) {
-                    $brands[] = $brand;
-                }
+            if (0 >= count($matches)) {
+                continue;
             }
+
+            $brand = $matches[2];
+            if (empty($brand)) {
+                continue;
+            }
+
+            $brands[] = $brand;
         }
+
         $brands = array_unique($brands);
 
-        usort($brands, static function ($a, $b): int {
-            return mb_strlen($b) - mb_strlen($a);
-        });
+        usort($brands, static fn ($a, $b): int => mb_strlen($b) - mb_strlen($a));
 
         return $brands;
     }
